i guess i'll bump for now (see desc.)
I can probably very easily bodge functionality into what River has written for automation to update _a_ file (not necessarily this one) so I don't have to update it every time. This is the one thing left I can think of that can be automated; so it'd be fantastic if that can get done.